+++ Magic/src/main/java/com/elmakers/mine/bukkit/wand/WandTemplate.java
@@ -11,6 +11,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableSet;
 
 import org.bukkit.Location;
+import org.bukkit.attribute.Attribute;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
 
@@ -41,6 +42,29 @@ public class WandTemplate extends BaseMagicProperties implements com.elmakers.mi
 
     public WandTemplate(MageController controller, String key, ConfigurationSection node) {
         super(controller);
+
+        // Migrate attributes
+        attributes = node.getConfigurationSection("item_attributes");
+        ConfigurationSection migrateAttributes = node.getConfigurationSection("attributes");
+        if (migrateAttributes != null) {
+            Set<String> keys = migrateAttributes.getKeys(false);
+            for (String attributeKey : keys) {
+                try {
+                    Attribute attribute = Attribute.valueOf(attributeKey.toUpperCase());
+                    if (attribute != null) {
+                        if (attributes == null) {
+                            attributes = node.createSection("item_attributes");
+                        }
+                        attributes.set(attributeKey, migrateAttributes.get(attributeKey));
+                        node.set("attributes", null);
+                        controller.getLogger().warning("You have vanilla item attributes in the 'attributes' property of wand template '" + key + "', please rename that to item_attributes.");
+                    }
+                } catch (Exception ex) {
+
+                }
+            }
+        }
+
         this.load(node);
         this.key = key;
 
@@ -51,8 +75,7 @@ public class WandTemplate extends BaseMagicProperties implements com.elmakers.mi
         migrateIcon = node.getString("migrate_icon");
         restorable = node.getBoolean("restorable", true);
         icon = node.getString("icon");
-        attributes = node.getConfigurationSection("attributes");
-        attributeSlot = node.getString("attribute_slot");
+        attributeSlot = node.getString("item_attribute_slot", node.getString("attribute_slot"));
         parentKey = node.getString("inherit");
 
         // Remove some properties that should not transfer to wands